+++ src/app/page.tsx
@@ -13,11 +13,25 @@ import SearchControl from "@/components/search/SearchControl";
 import SurveyBanner from "@/components/Survey";
 import SearchWrapper from "@/components/search/SearchWrapper";
 
+// import axios from "axios";
+import Cookies from "js-cookie";
+import { jwtDecode } from "jwt-decode";
+
 export default function Home() {
   const [isLoading, setIsLoading] = useState(true);
   const [clubs, setClubs] = useState<IClub[]>([]);
   const [currentClubs, setCurrentClubs] = useState<IClub[]>([]);
   const [followedClubs, setFollowedClubs] = useState<string[]>([]);
+  const token = Cookies.get("token");
+  let netid = "";
+
+  if (token) {
+    try {
+      netid = jwtDecode<{ netid: string }>(token).netid;
+    } catch (error) {
+      console.error("Failed to decode token:", error);
+    }
+  }
 
   useEffect(() => {
     const fetchApiMessage = async () => {
@@ -34,6 +48,23 @@ export default function Home() {
 
     fetchApiMessage();
   }, [setFollowedClubs, setClubs, setIsLoading]);
+
+  useEffect(() => {
+    const fetchUser = async () => {
+      try {
+        // get /users&netid
+        const response = await axios.get("/api/users", {
+          params: { netid: netid },
+        });
+
+        setFollowedClubs(response.data.user.followedClubs);
+      } catch (error) {
+        console.error("Failed to fetch user data:", error);
+      }
+    };
+
+    fetchUser();
+  }, [netid]);
 
   return (
     <AuthWrapper>
@@ -52,7 +83,12 @@ export default function Home() {
                 followedClubs={followedClubs}
               />
             </SearchWrapper>
-            <Catalog clubs={currentClubs} isLoading={isLoading} />
+            <Catalog
+              clubs={currentClubs}
+              isLoading={isLoading}
+              followedClubs={followedClubs}
+              setFollowedClubs={setFollowedClubs}
+            />
             <Footer />
           </div>
         </section>
